requirements.txt
pydantic==2.4.2
langchain==0.0.316
streamlit_chat
streamlit
openai
pandasai
pydantic==2.4.2
langchain==0.0.316
accelerate
sentence_transformers
streamlit_chat
streamlit
faiss-cpu
tiktoken
huggingface-hub
pypdf
python-dotenv
replicate
docx2txt
text_generation
openai==0.28.1
PyMuPDF
Pillow
pytesseract
pytest-shutil